Moved OptionChain class to own file as apart of the tda namespace
